Switch to github codeload & REL1_35 hashes
This mediawiki app is currently using 1.35 (even though we do
want to move to 1.36), so for now we need to use the 1.35 HEAD.

Also switch from extdist to github code load for consistency.
Also extdist tars actually disappear every time a new commit gets
pushed to the extension branch, which is annoying for reproducability.

I should write some docs for adding new extensions to this repo!