Don't ignore composer platform reqs (#147)
`--ignore-platform-reqs` was added here because composer refused to
update due to missing php extensions in the docker image. This problem
went away since the composer docker image we use was updated in
4a0e41b30ed9e80ec18cccb3a6af0e3f2e3dc961.

Having `--ignore-platform-reqs` now caused a problem because it means
that it also ignores the PHP version, and running `wbstack/sync.sh` on
the current main branch also pulls in PHP 8 packages which can't be used
yet.