Use port 80 for mediawiki in docker-compose
This was removed in https://github.com/wbstack/mediawiki/pull/136
But as noted in that PR it was not known what might break.
It turns out mediawiki redirects post saving are one example of
a broken thing.

The easiest fix for now, with the given uncertainty of this
docker-compose setup anyway is probably to use port 80.