Add a triggerable workflow for running and commiting sync
Not tested yet, but in principle the sync already works, as
we have a job that runs a diff to disallow merging.
So as long as the commit action works this should work.

I wrote this as I am using mobile data and don't want to run
it all locally...